move settings tracking stuff into its own module
